+++ src/RecoveryServices/RecoveryServices.SiteRecovery/ReplicationProtectedItem/SetAzureRmRecoveryServicesAsrReplicationProtectedItem.cs
@@ -130,6 +130,12 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
         [Parameter]
         public string RecoveryAvailabilitySet { get; set; }
 
+        // <summary>
+        ///     Gets or sets the resource ID of the availability zone to failover this virtual machine to.
+        /// </summary>
+        [Parameter]
+        public string RecoveryAvailabilityZone { get; set; }
+
         /// <summary>
         ///     Gets or sets the proximity placement group Id for replication protected item after failover.
         /// </summary>
@@ -277,6 +283,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
                     string.IsNullOrEmpty(this.PrimaryNic) &&
                     this.UseManagedDisk == null &&
                     this.IsParameterBound(c => c.RecoveryAvailabilitySet) &&
+                    this.IsParameterBound(c => c.RecoveryAvailabilityZone) &&
                     this.IsParameterBound(c => c.RecoveryProximityPlacementGroupId) &&
                     string.IsNullOrEmpty(this.RecoveryCloudServiceId) &&
                     string.IsNullOrEmpty(this.RecoveryResourceGroupId) &&
@@ -332,6 +339,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
                 var useManagedDisk = this.UseManagedDisk;
                 var availabilitySetId = this.RecoveryAvailabilitySet;
                 var proximityPlacementGroupId = this.RecoveryProximityPlacementGroupId;
+                var availabilityZone = this.RecoveryAvailabilityZone;
                 var primaryNic = this.PrimaryNic;
                 var diskIdToDiskEncryptionMap = this.DiskIdToDiskEncryptionSetMap;
                 var tfoNetworkId = string.Empty;
@@ -368,6 +376,14 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
                     availabilitySetId = this.IsParameterBound(c => c.RecoveryAvailabilitySet)
                         ? this.RecoveryAvailabilitySet
                         : providerSpecificDetails.RecoveryAvailabilitySetId;
+
+                    availabilityZone = this.IsParameterBound(c => c.RecoveryAvailabilityZone)
+                        ? this.RecoveryAvailabilityZone
+                        : providerSpecificDetails.TargetAvailabilityZone;
+
+                    proximityPlacementGroupId = this.IsParameterBound(c => c.RecoveryProximityPlacementGroupId)
+                        ? this.RecoveryProximityPlacementGroupId
+                        : providerSpecificDetails.TargetProximityPlacementGroupId;
 
                     if (string.IsNullOrEmpty(this.UseManagedDisk))
                     {
@@ -414,7 +430,9 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
                                 RecoveryAzureV1ResourceGroupId = null,
                                 RecoveryAzureV2ResourceGroupId = recoveryResourceGroupId,
                                 UseManagedDisks = useManagedDisk,
-                                DiskIdToDiskEncryptionMap = this.DiskIdToDiskEncryptionSetMap
+                                DiskIdToDiskEncryptionMap = this.DiskIdToDiskEncryptionSetMap,
+                                TargetAvailabilityZone = availabilityZone,
+                                TargetProximityPlacementGroupId = proximityPlacementGroupId
                             };
                     }
 
@@ -447,6 +465,12 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
 
                     availabilitySetId = this.IsParameterBound(c => c.RecoveryAvailabilitySet)
                         ? this.RecoveryAvailabilitySet : providerSpecificDetails.RecoveryAvailabilitySetId;
+
+                    availabilityZone = this.IsParameterBound(c => c.RecoveryAvailabilityZone)
+                        ? this.RecoveryAvailabilityZone : providerSpecificDetails.TargetAvailabilityZone;
+
+                    proximityPlacementGroupId = this.IsParameterBound(c => c.RecoveryProximityPlacementGroupId)
+                        ? this.RecoveryProximityPlacementGroupId : providerSpecificDetails.TargetProximityPlacementGroupId;
 
                     if (string.IsNullOrEmpty(this.UseManagedDisk))
                     {
@@ -484,7 +508,9 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
                             {
                                 RecoveryAzureV1ResourceGroupId = null,
                                 RecoveryAzureV2ResourceGroupId = recoveryResourceGroupId,
-                                UseManagedDisks = useManagedDisk
+                                UseManagedDisks = useManagedDisk,
+                                TargetAvailabilityZone = availabilityZone,
+                                TargetProximityPlacementGroupId = proximityPlacementGroupId
                             };
                     }
                     vMNicInputDetailsList = getNicListToUpdate(providerSpecificDetails.VmNics);