+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Introduced NetworkAclBypass and NetworkAclBypassResourceIds for Database Account cmdlets.
 
 ## Version 1.0.0
 * General availability of 'Az.CosmosDB' module