+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,29 @@
 --->
 
 ## Upcoming Release
+* Registered 8 new privatelink configurations
+    - `Microsoft.Automation/automationAccounts`
+    - `Microsoft.Compute/diskAccesses`
+    - `Microsoft.EventHub/namespaces`
+    - `Microsoft.Media/mediaservices`
+    - `Microsoft.Search/searchServices`
+    - `Microsoft.ServiceBus/namespaces`
+    - `Microsoft.Synapse/workspaces`
+    - `Microsoft.Purview/accounts`
+* Updated the privatelink configurations of 'Microsoft.Keyvault/vaults' and 'Microsoft.Web/sites' to higher quality ones
+
+## Version 4.6.0
+* Added new cmdlets to replace old product name `virtual router` with new name `route server` in the future.
+    - `New-AzRouteServer`
+    - `Get-AzRouteServer`
+    - `Remove-AzRouteServer`
+    - `Update-AzRouteServer`
+    - `Get-AzRouteServerPeer`
+    - `Add-AzRouteServerPeer`
+    - `Update-AzRouteServerPeer`
+    - `Remove-AzRouteServerPeer`
+    - Added deprecation attribute warning to the old cmdlets.
+* Bug fix in ExpressRouteLink MacSecConfig. Added new property `SciState` to `PSExpressRouteLinkMacSecConfig`
 * Updated format list and format table views for Get-AzVirtualNetworkGatewayConnectionIkeSa
 * Updated `set-azExpressRouteGateway` to allow parameter -MinScaleUnits without specifying -MaxScaleUnits
 * Updated cmdlets to enable setting of VpnLinkConnectionMode on VpnSiteLinkConnections.